Fix bug involving first-run of fetch-ipranges.sh
During the first run, it is impossible to back up the existing allowed-cidrs.txt file (because it does not exist!), therefore it fails with `mv: cannot stat '/root/allowed-cidrs.txt': No such file or directory`. Do not attempt to back it up on the first run.

+++ roles/denbi.osiris/files/fetch-ipranges.sh
@@ -8,7 +8,7 @@ NEW="/root/allowed-cidrs.txt.new"
 wget --quiet https://www.ipdeny.com/ipblocks/data/aggregated/de-aggregated.zone -O "$NEW"
 
 if grep 194.94 "$NEW" > /dev/null ; then
-    mv -f "$NOW" "$OLD"
+    [ -f "$NOW" ] && mv -f "$NOW" "$OLD"
     mv -f "$NEW" "$NOW"
 else
     echo "Expected IP range not found. Keeping previous ranges file"
